requirements.txt
streamlit
ultralytics
opencv-python
Pillow
torch
streamlit-webrtc
fastapi
uvicorn[standard]
websockets